+++ R/connection.R
@@ -121,58 +121,111 @@ close_connection <- function(conn) {
 #' @param allow_table_only
 #'  logical. If `TRUE`, allows for returning an `DBI::Id` with `table` = `myschema.table` if schema `myschema`
 #'  is not found in `conn`.
-#' @details The given db_table_id is parsed using an assumption of "schema.table" syntax into
-#'  a DBI::Id object with corresponding schema (if the conn supports it) and table values.
-#' @return A DBI::Id object parsed from db_table_id
+#'  If `FALSE`, the function will raise an error if the implied schema cannot be found in `conn`
+#' @details The given `db_table_id` is parsed to a DBI::Id depending on the type of input:
+#'  * `character`: db_table_id is parsed to a DBI::Id object using an assumption of "schema.table" syntax
+#'     with corresponding schema (if found in `conn`) and table values.
+#'     If no schema is implied, the default schema of `conn` will be used.
+#'
+#'  * `DBI::Id`: if schema is not specified in `Id`, the schema is set to the default schema for `conn` (if given).
+#'
+#'  * `tbl_sql`: the remote name is used to resolve the table identification.
+#'
+#' @return A DBI::Id object parsed from db_table_id (see details)
 #' @examples
 #' id("schema.table")
 #' @seealso [DBI::Id] which this function wraps.
 #' @export
 id <- function(db_table_id, conn = NULL, allow_table_only = TRUE) {
-  if (inherits(db_table_id, "Id")) {
-    return(db_table_id)
-  }
-
-  if (!allow_table_only) checkmate::check_class(conn, "DBIConnection")
-
   UseMethod("id")
 }
 
+
+#' @export
+id.Id <- function(db_table_id, conn = NULL, allow_table_only = TRUE) {
+  return(DBI::Id(schema = purrr::pluck(db_table_id, "name", "schema", .default = SCDB::get_schema(conn)),
+                 table = purrr::pluck(db_table_id, "name", "table")))
+}
+
+
 #' @export
 id.character <- function(db_table_id, conn = NULL, allow_table_only = TRUE) {
+
+  checkmate::assert(is.null(conn), DBI::dbIsValid(conn), combine = "or")
 
   if (stringr::str_detect(db_table_id, "\\.")) {
     db_name <- stringr::str_split_1(db_table_id, "\\.")
     db_schema <- db_name[1]
     db_table  <- db_name[2]
 
-    # If no matching implied schema is found, return the unmodified db_table_id
+    # If no matching implied schema is found, return the unmodified db_table_id in the default schema
     if (allow_table_only && !is.null(conn) && !schema_exists(conn, db_schema)) {
-      return(DBI::Id(table = db_table_id))
+      return(DBI::Id(schema = get_schema(conn), table = db_table_id))
     }
   } else {
-    db_schema <- NULL
+    db_schema <- get_schema(conn)
     db_table <- db_table_id
   }
 
   return(DBI::Id(schema = db_schema, table = db_table))
 }
 
+
 #' @export
 id.tbl_dbi <- function(db_table_id, conn = NULL, allow_table_only = TRUE) {
-  table_ident <- dbplyr::remote_table(db_table_id)
 
-  id <- with(table_ident, {
-    list(catalog = catalog,
-         schema = schema,
-         table = table) |>
-      (\(.x) subset(.x, !is.na(.x)))() |>
-      do.call(DBI::Id, args = _)
-  })
+  # If table identification is fully qualified extract Id from remote_Table
+  if (!is.na(purrr::pluck(dbplyr::remote_table(db_table_id), unclass, "schema"))) {
 
-  if (!is.null(conn) && !identical(conn, dbplyr::remote_con(db_table_id))) {
-    rlang::warn("Table connection is different than conn")
+    table_ident <- dbplyr::remote_table(db_table_id) |>
+      unclass() |>
+      purrr::discard(is.na)
+
+    table_conn <- dbplyr::remote_con(db_table_id)
+
+    return(
+      DBI::Id(
+        catalog = purrr::pluck(table_ident, "catalog"),
+        schema = purrr::pluck(table_ident, "schema", .default = get_schema(table_conn)),
+        table = purrr::pluck(table_ident, "table")
+      )
+    )
+
+  } else {
+
+    # If not attempt to resolve the table from existing tables.
+    # For SQLite, there should only be one table in main/temp matching the table
+    # In some cases, tables may have been added to the DB that makes the id ambiguous.
+    schema <- get_tables(dbplyr::remote_con(db_table_id), show_temporary = TRUE) |>
+      dplyr::filter(.data$table == dbplyr::remote_name(db_table_id)) |>
+      dplyr::pull("schema")
+
+    if (length(schema) > 1)  {
+      stop(
+        "Table identification has been corrupted! ",
+        "This table does not contain information about its schema and ",
+        "multiple tables with this name were found across schemas."
+      )
+    }
+
+    return(DBI::Id(schema = schema, table = dbplyr::remote_name(db_table_id)))
   }
+}
 
-  return(id)
+
+#' @export
+as.character.Id <- function(x, ...) {
+
+  info <- x@name |>
+    purrr::discard(is.na)
+
+  id_representation <- list(
+    catalog = purrr::pluck(info, "catalog"),
+    schema = purrr::pluck(info, "schema"),
+    table = purrr::pluck(info, "table")
+  ) |>
+    purrr::discard(is.null) |>
+    do.call(purrr::partial(paste, sep = "."), args = _)
+
+  return(id_representation)
 }