+++ NEWS.md
@@ -1,17 +1,56 @@
 # SCDB (development version)
 
-## Minor Improvements and Fixes
+## BREAKING CHANGES:
+
+* Table identification is now more specific (#93):
+
+  Most SCDB functions allow for tables to be specified by a character representation of "schema.table".
+
+  Before, if no schema was implied in this context, SCDB would attempt to match the table among both
+  permanent and temporary tables.
+
+  Now, it will always assume that a lack of schema means the default schema should be used.
+  This is also the case if `DBI::Id()` is used without a schema specification.
+
+* The `show_temporary` argument of `get_tables()` is now a simple logical (#93).
+
+  In addition, schema is always returned in the list of table (no longer NA for default schema).
+
+* Tables created with `create_table()` will now be temporary or permanent dependent on the default value of
+  `DBI::dbCreateTable()` (#93).
+
+  If you wish to overwrite this, use `...` arguments which are passed to `DBI::dbCreateTable()`.
+
+* If a `SQLiteConnection` is passed to `get_schema()`, the returned schema will always be "main" (#93).
+
+## Features
+* The S3 method `as.character.Id()` is added which converts `DBI::Id()` to `character` (#93).
+
+## Improvements and Fixes
+
+* Improvements for `create_table()` (#93):
+  - now writes the table if a remote connection is given. Before, it would only create the
+  table with corresponding columns.
+  - can now create temporary tables for Microsoft SQL Server.
 
 * Improved checks on `get_connection()` (#83):
   - If given, `host` does not need to look like an IP address (e.g. "localhost" is not unrealistic).
   - A `character` input for `port` is allowed if it is a string of digits.
   - Now checks if `timezone` and `timezone_out` is an IANA time zone.
 
-* `get_connection()` now checks the value of any `timezone` and `timezone_out` arguments.
+* `get_connection()` now checks the value of any `timezone` and `timezone_out` arguments (#83).
 
 * `table_exists()` now correctly gives ambiguity warning on Microsoft SQL Server and PostgreSQL backends (#80).
 
+* `get_tables()` now supports temporary tables for Microsoft SQL Server (#93).
+
 * Fixed dplyr joins failing if `testthat` is not installed (#90).
+
+## Testing
+
+Added missing tests for `create_logs_if_missing()` (#93).
+
+Improved tests for `get_tables()`, `table_exists()`, and `create_table()` (#93).
 
 # SCDB 0.3
 
@@ -27,9 +66,9 @@
 
 * Fixed `update_snapshot()` not working with a `DBI::Id()` instance as `db_table` argument (#72).
 
-* Suppressed recurring messages from dbplyr >= 2.4.0 about table names containing `.`.
+* Suppressed recurring messages from dbplyr >= 2.4.0 about table names containing `.` (#72).
 
-* Added `show_temp` option to `get_tables()` to allow retrieving temporary tables.
+* Added `show_temp` option to `get_tables()` to allow retrieving temporary tables (#72).
 
 ## Other news